+++ org.lxtk/src/org/lxtk/DocumentSemanticTokensProvider.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 1C-Soft LLC.
+ * Copyright (c) 2021, 2022 1C-Soft LLC.
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which is available at
@@ -77,11 +77,10 @@ public interface DocumentSemanticTokensProvider
         SemanticTokensRangeParams params);
 
     /**
-     * Returns a stream of events that are emitted when semantic tokens from this provider
-     * have changed.
+     * Returns a stream of events that are emitted when semantic tokens need to be refreshed.
      *
-     * @return a stream of events that are emitted when semantic tokens from this provider
-     *  have changed (never <code>null</code>)
+     * @return a stream of events that are emitted when semantic tokens need to be refreshed
+     *  (never <code>null</code>)
      */
-    EventStream<Void> onDidChangeSemanticTokens();
+    EventStream<Void> onRefreshSemanticTokens();
 }